requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Werkzeug==2.3.7
Jinja2==3.1.2
python-dotenv==1.0.0
Pillow==10.0.1
python-magic==0.4.27
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
xgboost==1.7.6
easyocr==1.7.0
pytesseract==0.3.10
opencv-python==4.8.1.78
networkx==3.1
python-louvain==0.16
shap==0.42.1
lime==0.2.0.1
matplotlib==3.7.2
seaborn==0.12.2
gunicorn==21.2.0
alembic==1.12.0
SQLAlchemy==2.0.21
WTForms==3.0.1
email-validator==2.0.0
